[fix] アラートの表示のタスク名が適切でない不具合を修正
