+++ RepeatReminder/RepeatReminder/Presentation/Views/TaskListView.swift
@@ -9,7 +9,6 @@ import SwiftUI
 
 struct TaskListView: View {
     @ObservedObject var viewModel = TaskListViewModel()
-    @State private var isShowedAlert = false
     
     var body: some View {
         NavigationStack{
@@ -84,34 +83,7 @@ struct TaskListView: View {
                                     // タスクの情報を表示
                                     TaskCell(task:task)
                                     // ボタンを表示
-                                    VStack(spacing:4){
-                                        NavigationLink(
-                                            destination: TaskAddEditView(isEditing:true,task:task)
-                                        ){
-                                            Image(systemName:"pencil.circle")
-                                                .foregroundColor(Color("ButtonColor"))
-                                                .font(.system(size:48))
-                                        }
-                                        Button {
-                                            print("tap complete button")
-                                            isShowedAlert = true
-                                        } label: {
-                                            Image(systemName:"checkmark.circle.fill")
-                                                .foregroundColor(Color("ButtonColor"))
-                                                .font(.system(size:48))
-                                        }
-                                        .alert("このタスクを完了しますか？\n"+task.name, isPresented: $isShowedAlert) {
-                                            Button("キャンセル",role:.cancel) {
-                                                isShowedAlert = false
-                                            }
-                                            Button("OK") {
-                                                isShowedAlert = false
-                                                viewModel.completeTask(task: task)
-                                            }
-                                        } message: {
-                                            Text("設定から復元できます")
-                                        }
-                                    }
+                                    ClearButton(task:task,viewModel:viewModel)
                                 }
                             }
                         }.padding(.leading,8)
